Update to check for radarid to be more general script.

+++ site-linux/rsync_to_sas_from_nas.sh
@@ -15,7 +15,14 @@ source $HOME/.bashrc # source the RADARID, SDCOPY and other things
 #HOME=/home/radar/
 DMAP_SOURCE=/borealis_nfs/borealis_data/rawacf_dmap/
 ARRAY_SOURCE=/borealis_nfs/borealis_data/rawacf_array/
-DEST=/data/${RADARID}_data/
+
+# CLY - move to holding dir to do dmaps on SDCOPY
+if [[ "${RADARID}" == "cly" ]]; then
+	DEST=/sddata/cly_holding_dir
+else
+	DEST=/data/${RADARID}_data/
+fi
+
 
 echo ""
 date
@@ -30,22 +37,26 @@ if [[ "$RSYNCRUNNING" == *"data_flow/site-linux/rsync_to_sas_from_nas"*"data_flo
   exit
 fi
 
-files=`find ${DMAP_SOURCE} -name '*rawacf.bz2' -printf '%p\n'`
-for file in $files
-do
-        datafile=`basename $file`
-        path=`dirname $file`
-        cd $path
-        rsync -av --partial --partial-dir=${TEMPDEST} --timeout=180 --rsh=ssh ${datafile} ${SDCOPY}:${DEST}
-        # check if transfer was okay using the md5sum program
-        ssh ${SDCOPY} "cd ${DEST}; md5sum -b ${datafile}" > ${MD5}
-        md5sum -c ${MD5}
-        mdstat=$?
-        if [ ${mdstat} -eq 0 ] ; then
-                echo "Deleting file: "${file}
-                rm -v ${file}
-        fi
-done
+if [[ "${RADARID}" == "cly" ]]; then
+	echo "Not transferring any dmaps, only HDF5"
+else
+	files=`find ${DMAP_SOURCE} -name '*rawacf.bz2' -printf '%p\n'`
+	for file in $files
+	do
+		datafile=`basename $file`
+		path=`dirname $file`
+		cd $path
+		rsync -av --partial --partial-dir=${TEMPDEST} --timeout=180 --rsh=ssh ${datafile} ${SDCOPY}:${DEST}
+		# check if transfer was okay using the md5sum program
+		ssh ${SDCOPY} "cd ${DEST}; md5sum -b ${datafile}" > ${MD5}
+		md5sum -c ${MD5}
+		mdstat=$?
+		if [ ${mdstat} -eq 0 ] ; then
+                	echo "Deleting file: "${file}
+                	rm -v ${file}
+        	fi
+	done
+fi
 
 files=`find ${ARRAY_SOURCE} -name '*rawacf.hdf5' -printf '%p\n'`
 for file in $files
